simulation/src/vehicle/obu/requirements.txt
environs==9.5.0
asyncio-dgram==2.1.2
aiohttp==3.8.1
PyJWT==2.4.0